cloud_function/requirements.txt
google-cloud-bigquery==3.11.4
functions-framework==3.*